requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
pyTelegramBotAPI==4.29.1
requests==2.32.5
urllib3==2.5.0
python-dotenv==0.21.1
Django==3.2.25
django-filter==23.5
djangorestframework==3.15.1